deployment/frontend/requirements.txt
requests
streamlit
typing_extensions
Pillow